test: [WIP] Testing changing type inside structs

+++ demon.h
@@ -1,9 +1,6 @@
 #ifndef DEMON_H
 #define DEMON_H
 
-#include <stdio.h>
-#include <stdlib.h>
-
 #define STEP_SIZE 5
 #define BLOCK_SIZE 5
 #define SEARCH_RADIUS 10
@@ -11,8 +8,8 @@
 
 typedef struct
 {
-    int width;
-    int height;
+    size_t width;
+    size_t height;
     float *data; // in [0,1]
 } Image;
 
@@ -20,11 +17,11 @@ typedef struct
 {
     // displacement field is computed at 1/stepSize resolution and interpolated
     // to scale
-    int width;
-    int height;
+    size_t width;
+    size_t height;
     float *x;
     float *y;
-    int stepSize; // in pixels
+    size_t stepSize; // in pixels
 } DispField;
 
 typedef struct
@@ -34,14 +31,14 @@ typedef struct
 } DispVect; // in pixels
 
 // Initialization
-DispField *initDispField(int width, int height, int stepSize);
+DispField *initDispField(size_t width, size_t height, size_t stepSize);
 
 // Preprocessing
 void normalizeImage(Image *img);
 
 // Block comparison
-float compareBlockSSD(Image *fixed, Image *moving, int fixedX, int fixedY,
-                      int movingX, int movingY, int blockSize);
+float compareBlockSSD(Image *fixed, Image *moving, size_t fixedX, size_t fixedY,
+                      size_t movingX, size_t movingY, size_t blockSize);
 
 // Disp estimation
 void estimateBlockDisps(Image *fixed, Image *moving, DispField *df,
